BUGFIX: By adding items from some games (such as Spiral Knights), user can consistently cause the bot to crash.  THIS IS A BREAKING CHANGE :( Also, unexpected side-fix:  Trade.GetItemFromPrivateBp() appears to have been incorrect.

+++ SteamBot/SteamTradeDemoHandler.cs
@@ -60,7 +60,7 @@ namespace SteamBot
         {
             Bot.SteamFriends.SetPersonaState(EPersonaState.Busy);
             // NEW -------------------------------------------------------------------------------
-            List<int> contextId = new List<int>();
+            List<long> contextId = new List<long>();
             tested = false;
 
             /*************************************************************************************
